requirements.txt
# ── Microsoft Agent Framework (MAF) ──────────────────────────
# Core + orchestrations + Ollama provider
agent-framework-core[all] @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/core
agent-framework @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python
agent-framework-ollama @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/ollama
agent-framework-orchestrations @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/orchestrations

# ── Configuration ────────────────────────────────────────────
python-dotenv
